hotfix

+++ math-service/src/main/scala/com/mathservice/application/VolatilityService.scala
@@ -5,7 +5,7 @@ import com.mathservice.domain.VolatilityCalculator
 import com.typesafe.scalalogging.LazyLogging
 
 class VolatilityService[F[_]: Sync] extends LazyLogging {
-  def calculateVolatility(returns: List[Float]): F[Either[String, Double]] = {
+  def calculateVolatility(returns: List[Double]): F[Either[String, Double]] = {
     Sync[F].delay {
       logger.info(s"Calculating volatility for ${returns.length} data points")
       VolatilityCalculator.calculateVolatility(returns)
